pyproject.toml
[project]
name = "kursach-checker"
version = "0.1.0"
description = "Система технической проверки документов: Telegram бот + Mini App + Админка"
requires-python = ">=3.11"
dependencies = [
  "fastapi==0.116.1",
  "uvicorn[standard]==0.35.0",
  "sqlalchemy[asyncio]==2.0.43",
  "asyncpg==0.30.0",
  "redis==6.4.0",
  "arq==0.26.3",
  "pydantic==2.11.7",
  "pydantic-settings==2.10.1",
  "python-multipart==0.0.20",
  "python-docx==1.2.0",
  "aiogram==3.22.0",
  "httpx==0.28.1",
  "PyJWT==2.10.1",
  "passlib[bcrypt]==1.7.4",
  "alembic==1.16.5",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]
